scripts/requirements_telethon.txt
# Requirements for Telethon whitelist seeding script
telethon>=1.34.0
asyncpg>=0.29.0
python-dotenv>=1.0.0